requirements.txt
uvicorn
fastapi
fasttext
pydantic
websocket-client
qdrant-client
prometheus_client
redis
numpy==1.26.4
pandas
google-genai
python-dotenv
requests
geopy
geocoder
urllib3
certifi
boto3
groq
deepl
pdfplumber
python-docx
prometheus_fastapi_instrumentator
pymysql
apscheduler